% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocMatrix.R
\name{assocMatrix}
\alias{assocMatrix}
\title{Creating an Association Matrix}
\usage{
assocMatrix(dat, n.levels = 6, plotTitle = "default",
  val.label = FALSE)
}
\arguments{
\item{dat}{Dataframe with at least 2 categorical variables}

\item{n.levels}{Specifying the number of levels for all categorical variables to be considered (default is 6)}

\item{plotTitle}{Optional string specifying the title of the plot}

\item{val.label}{Boolean determining whether to add values of coefficients to the plot (default is FALSE)}
}
\value{
Heat map style plot displaying the strength of pair-wise associations in categorical variables
}
\description{
Using Cramer's V, the association matrix shows pair-wise relationships between categorical variables with up to a specified number of levels
}
\examples{
assocMatrix(mtcars)
}
