% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrFinder.R
\name{corrFinder}
\alias{corrFinder}
\title{Finding pairwise and average correlations}
\usage{
corrFinder(dat, strongest = TRUE, weakest = FALSE)
}
\arguments{
\item{dat}{Dataframe with at least two numeric columns}

\item{strongest}{Boolean indicating whether to order the output by the strongest absolute correlations (default is TRUE)}

\item{weakest}{Boolean indicating whether to order the output by the weakest absolute correlations (default is FALSE)}
}
\value{
Returns a dataframe storing all pairwise Pearson correlation coefficients and corresponding p-values, as well as average correlation by variable
}
\description{
Finding pairwise and average correlations
}
\examples{
corrFinder(iris)
}
