% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMatrix.R
\name{corrMatrix}
\alias{corrMatrix}
\title{Correlation Matrix}
\usage{
corrMatrix(dat, plotTitle = "default", val.label = FALSE, output = "plot")
}
\arguments{
\item{dat}{Dataframe with at least 2 numeric variables}

\item{plotTitle}{Optional string specifying the title of the plot}

\item{val.label}{Boolean determining whether to add values of coefficients to the plot (default is FALSE)}

\item{output}{Specifying whether the output is to be a plot of raw data (default is plot)}
}
\value{
Heat map style plot displaying the strength of pair-wise correlations in numeric variables
}
\description{
Correlation Matrix
}
\examples{
corrMatrix(iris)
}
