% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_options.R
\name{data_options}
\alias{data_options}
\title{Data Options}
\usage{
data_options()
}
\value{
List of datasets
}
\description{
Function showing which datasets are supported by the `read_data` function.
}
\examples{

data_options()

}
