% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{expose}
\alias{expose}
\title{expose}
\usage{
expose(dat, target = NA, n.levels = 6)
}
\arguments{
\item{dat}{Dataframe object}

\item{target}{Name of column from datafame for uncovering insights. Must be a string.}

\item{n.levels}{Specified number of levels to treat a factor variable (default is six)}
}
\value{
Insights report based on input target
}
\description{
expose
}
