% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_ghd}
\alias{fit_ghd}
\alias{fit_ghd.default}
\alias{fit_ghd.tbl}
\alias{fit_ghd.xts}
\alias{fit_ghd.matrix}
\alias{fit_ghd.numeric}
\title{Estimation of the Generalized Hyperbolic Distribution}
\usage{
fit_ghd(.invariant, .symmetric = FALSE)

\method{fit_ghd}{default}(.invariant, .symmetric = FALSE)

\method{fit_ghd}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_ghd}{xts}(.invariant, .symmetric = FALSE)

\method{fit_ghd}{matrix}(.invariant, .symmetric = FALSE)

\method{fit_ghd}{numeric}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular data structure.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
An object of the the class \code{mle.ghyp}.
}
\description{
This function is a wrapper around the \code{\link[ghyp]{fit.ghypuv}} and
\code{\link[ghyp]{fit.ghypmv}}, that performs a maximum likelihood estimation
on generalized hyperbolic distribution.
}
\examples{
##### fit_ghd #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_ghd(x) # multivariate estimation

fit_ghd(x[ , 3, drop = FALSE]) # univariate estimation
}
\seealso{
\code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_vg}} \code{\link{fit_t}}
}
