% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_hyp}
\alias{fit_hyp}
\alias{fit_hyp.default}
\alias{fit_hyp.tbl}
\alias{fit_hyp.xts}
\alias{fit_hyp.matrix}
\alias{fit_hyp.numeric}
\title{Estimation of the Hyperbolic Distribution}
\usage{
fit_hyp(.invariant, .symmetric = FALSE)

\method{fit_hyp}{default}(.invariant, .symmetric = FALSE)

\method{fit_hyp}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_hyp}{xts}(.invariant, .symmetric = FALSE)

\method{fit_hyp}{matrix}(.invariant, .symmetric = FALSE)

\method{fit_hyp}{numeric}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) data structure.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
An object of the the class \code{uncover_fit}.
}
\description{
This function performs a maximum likelihood estimation
on the hyperbolic distribution.
}
\examples{
##### fit_hyp #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_hyp(x) # multivariate estimation

fit_hyp(x[ , 4]) # univariate estimation
}
\seealso{
\code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_vg}} \code{\link{fit_t}}
}
