% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_nig}
\alias{fit_nig}
\alias{fit_nig.default}
\alias{fit_nig.tbl}
\alias{fit_nig.xts}
\alias{fit_nig.matrix}
\alias{fit_nig.numeric}
\title{Estimation of the Normal-Inverse Gaussian Distribution}
\usage{
fit_nig(.invariant, .symmetric = FALSE)

\method{fit_nig}{default}(.invariant, .symmetric = FALSE)

\method{fit_nig}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_nig}{xts}(.invariant, .symmetric = FALSE)

\method{fit_nig}{matrix}(.invariant, .symmetric = FALSE)

\method{fit_nig}{numeric}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular data structure.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
An object of the the class \code{mle.ghyp}.
}
\description{
This function is a wrapper around the \code{\link[ghyp]{fit.NIGuv}} and
\code{\link[ghyp]{fit.NIGmv}}, that performs a maximum likelihood estimation
on the Normal-Inverse Gaussian (NIG) distribution.
}
\examples{
##### fit_nig #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_nig(x) # multivariate estimation

fit_nig(x[ , 4, drop = FALSE]) # univariate estimation
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_vg}} \code{\link{fit_t}}
}
