% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_t}
\alias{fit_t}
\alias{fit_t.default}
\alias{fit_t.tbl}
\alias{fit_t.xts}
\alias{fit_t.matrix}
\alias{fit_t.numeric}
\title{Estimation of the Student-t Distribution}
\usage{
fit_t(.invariant, .symmetric = FALSE)

\method{fit_t}{default}(.invariant, .symmetric = FALSE)

\method{fit_t}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_t}{xts}(.invariant, .symmetric = FALSE)

\method{fit_t}{matrix}(.invariant, .symmetric = FALSE)

\method{fit_t}{numeric}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular data structure.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
An object of the the class \code{mle.ghyp}.
}
\description{
This function is a wrapper around the \code{\link[ghyp]{fit.tuv}} and
\code{\link[ghyp]{fit.tmv}}, that performs a maximum likelihood estimation
on the Student-t distribution.
}
\examples{
##### fit_t #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_t(x) # multivariate estimation

fit_t(x[ , 4, drop = FALSE]) # univariate estimation
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_vg}}
}
