% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_vg}
\alias{fit_vg}
\alias{fit_vg.default}
\alias{fit_vg.tbl}
\alias{fit_vg.xts}
\alias{fit_vg.matrix}
\alias{fit_vg.numeric}
\title{Estimation of the Variance-Gamma Distribution}
\usage{
fit_vg(.invariant, .symmetric = FALSE)

\method{fit_vg}{default}(.invariant, .symmetric = FALSE)

\method{fit_vg}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_vg}{xts}(.invariant, .symmetric = FALSE)

\method{fit_vg}{matrix}(.invariant, .symmetric = FALSE)

\method{fit_vg}{numeric}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) data structure.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
An object of the the class \code{uncover_fit}.
}
\description{
This function performs a maximum likelihood estimation
on the Variance-Gamma (VG) distribution.
}
\examples{
##### fit_vg #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_vg(x) # multivariate estimation

fit_vg(x[ , 4]) # univariate estimation
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_t}}
}
