% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{getStats}
\usage{
getStats(dat)
}
\arguments{
\item{dat}{Dataframe}
}
\value{
Statistics on the inputted dataset
}
\description{
Gets standard statistics from a dataset
}
\examples{
getStats(mtcars)
}
