% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gupi.R
\name{gupi}
\alias{gupi}
\title{Guilty Until Proven Innocent}
\usage{
gupi(v)
}
\arguments{
\item{v}{vector of numeric values to assess whether outliers exist}
}
\value{
dataframe with two columns: `index` and `value`. The `index` column represents which row contains the outlier, while the `value` row contains the outlying value
}
\description{
Guilty Until Proven Innocent
}
\examples{
iris[gupi(iris$Sepal.Width)$index,]

}
