% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{lbe.gen}
\alias{lbe.gen}
\title{Log Bayesian evidence generator}
\usage{
lbe.gen(
  method = "SMC_BIC",
  thres = Inf,
  obs_mat,
  res_vec,
  p_num = 0,
  rpri = NULL,
  p_pdf = NULL,
  subs = NULL
)
}
\arguments{
\item{method}{Method to be used to estimate log(Z), can be one of `"SMC"`,
`"SMC_BIC"` or `"BIC"`}

\item{thres}{The threshold for which the number of observations needs to
exceed to consider using BIC as an estimator. Only applies if method
`"SMC_BIC"` is selected. Defaults to infinity if not specified.}

\item{obs_mat}{Covariate matrix}

\item{res_vec}{Binary response vector}

\item{p_num}{Number of samples of the prior used for the SMC sampler. Not
required if method `"BIC"` selected.}

\item{rpri}{Function to sample from the prior. Must only have two arguments,
`p_n` and `di` (Number of prior samples to generate and the number of
dimensions of a single sample respectively).}

\item{p_pdf}{Probability Density Function of the prior. Must only have two
arguments, `th` and `di` (a vector or matrix of regression coefficients
samples and the number of dimensions of a single sample respectively).}
}
\value{
An estimation of the log Bayesian evidence
}
\description{
Estimates the log Bayesian evidence (log(Z)) of a Bayesian
logistic regression model.
}
\details{
log(Z) is estimated using three possible methods:

1. `"SMC"`: Estimates log(Z) using an SMC sampler. See `IBIS.Z` for more
details.

2. `"SMC_BIC"`: Initially tries to estimate log(Z) using BIC but will revert
to using an SMC sampler if the data is linearly separable. Also reverts to
using an SMC sampler if the number of observations is below a certain
threshold. Recommended method and is set as default.

3. `"BIC"`: Estimates log(Z) with the Bayesian Information Criterion (BIC).
Does not require prior specification but only advisable if the number of
observations is large.
}
\examples{

# First we generate a covariate matrix `obs_mat` and binary response vector
# `res_vec`
CM <- matrix(rnorm(2000),1000,2)
rv <- sample(0:1,1000,replace=T)

# We can then estimate log(Z) using the `"BIC"` method
lbe.gen(method = "BIC",obs_mat = CM,res_vec = rv)

# If we assume the prior for the regression coefficients is a standard
# normal
pr_samp <- function(p_n,di){return(rmvn(p_n,rep(0,di),diag(di)))}
pr_fun <- function(th,di){return(dmvn(th,mu=rep(0,di),sigma=diag(di)))}

# If we suspect that the data might be linearly separable
lbe.gen(method = "SMC_BIC",obs_mat = CM,res_vec = rv,p_num = 500, rpri = pr_samp, p_pdf = pr_fun)

# For a much smaller dataset
CM.2 <- matrix(rnorm(20),10,2)
rv.2 <- sample(0:1,10,replace=T)

# it is more preferable to use method `"SMC"`
lbe.gen(method = "SMC",obs_mat = CM.2,res_vec = rv.2,p_num = 500, rpri = pr_samp, p_pdf = pr_fun)

# For small datasets the estimates of log(Z) using BIC or SMC will differ
# significantly. This becomes relatively less problematic as the dataset
# becomes larger however.
sapply(10:11,FUN = function(u,covm,resv,ps,pf){lbe.gen(method="SMC_BIC",thres=u,obs_mat=covm,res_vec=resv,p_num=500,rpri=ps,p_pdf=pf)},covm = CM.2,resv = rv.2,ps = pr_samp,pf = pr_fun)
sapply(1000:1001,FUN = function(u,covm,resv,ps,pf){lbe.gen(method="SMC_BIC",thres=u,obs_mat=covm,res_vec=resv,p_num=500,rpri=ps,p_pdf=pf)},covm = CM,resv = rv,ps = pr_samp,pf = pr_fun)

}
\seealso{
[IBIS.Z]
}
\keyword{carlo}
\keyword{monte}
\keyword{sequential}
