% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{memo.bic}
\alias{memo.bic}
\title{Log Bayesian evidence estimator through BIC, built for memoisation}
\usage{
memo.bic(X, y, which_obs = 1:length(y), param_start = NULL)
}
\arguments{
\item{X}{Covariate matrix}

\item{y}{Binary response vector}

\item{which_obs}{The indices of the observations in `X` that are in the
model. Defaults to all observations if not specified.}

\item{param_start}{Starting values for the model parameters used in the
`glm` function.}
}
\value{
A list containing; logZ - An estimation of the log Bayesian
evidence, coeffs - The coefficients produced by the glm model required for
calculation of the BIC value and input - the input `which_obs`
}
\description{
Estimates the log Bayesian evidence (log(Z)) of a Bayesian
logistic regression model using BIC.
}
\details{
This function is built with memoisation in mind and is intended to
be used as a supplementary function to `lbe.gen`. Producing `which_obs` as
an output is used to obtain input information from keys in the cache when
reviewing inputs which have been memoised.
}
\examples{

# First we generate a covariate matrix `X` and binary response vector
# `y`
CM <- matrix(rnorm(2000000),1000000,2)
rv <- sample(0:1,1000000,replace=T)

# We can then estimate log(Z) using the `"BIC"` method
model.1000000 <- memo.bic(X = CM,y = rv)
model.1000000$logZ
model.1000000$coeffs

# If we are only concerned with the first 500000 observations
model.500000 <- memo.bic(X = CM,y = rv, which_obs = 1:500000)

# If we are then wanting the model to include the first 500001 observations,
# it may be faster to use the output of model.500000
system.time(memo.bic(X = CM,y = rv, which_obs = 1:500001))
system.time(memo.bic(X = CM,y = rv, which_obs = 1:500001,param_start = model.500000$coeffs))

}
\seealso{
[glm,BIC,lbe.gen]
}
\keyword{BIC}
\keyword{memoisation}
