% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noVarCols.R
\name{noVarCols}
\alias{noVarCols}
\title{noVarCols}
\usage{
noVarCols(dat)
}
\arguments{
\item{dat}{Dataframe}
}
\value{
Index positions for columns with no variance
}
\description{
Identifying all columns with no variance
}
\examples{
sub_iris  <- iris[1:50,]
noVarCols(sub_iris)
sub_iris <- sub_iris[,-noVarCols(sub_iris)]

}
