% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST.R
\name{one.stage.mst}
\alias{one.stage.mst}
\title{One stage minimum spanning tree generation}
\usage{
one.stage.mst(obs, rho = NULL)
}
\arguments{
\item{obs}{Covariate matrix}

\item{rho}{Vector of indices which indicate the variables of `obs` to use to construct the minimum spanning tree}
}
\value{
An `igraph` object of the minimum spanning tree induced subgraph
}
\description{
Generation of a minimum spanning tree induced subgraph using the covariate matrix provided.

Used in UNCOVER to generate the initial graph for the algorithm if the pruning method does not involve validation data.
}
\examples{

# First we generate a covariate matrix `obs`
CM <- matrix(rnorm(300),100,3)

# Assuming we require a minimum spanning tree induced subgraph using all
# variables
one.stage.mst(obs = CM)

# If we only require the minimum spanning tree to be constructed from the
# first two variables
one.stage.mst(obs = CM,rho = 1:2)

}
\keyword{minimum}
\keyword{spanning}
\keyword{tree}
