% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reading open-sourced data}
\usage{
read_data(open_dat)
}
\arguments{
\item{open_dat}{Character string specifying the dataset the user wishes to read. Use `uncoveR::data_options()` to see the dataset options available.}
}
\value{
Dataframe object containing the data specified.
}
\description{
This function is set up to pull open sourced datasets from the web.
}
\examples{
healthDat <- read_data(open_dat = 'pophealth')
}
