% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bootstrap_scenarios}
\alias{ffp_moments}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ffp}{\code{\link[ffp]{bootstrap_scenarios}}, \code{\link[ffp]{ffp_moments}}}
}}

