% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal}
\alias{reveal}
\alias{reveal.default}
\alias{reveal.tbl_df}
\alias{reveal.xts}
\alias{reveal.matrix}
\title{Distribution Selection with AIC Criterion}
\usage{
reveal(.invariant, .silent = TRUE, ...)

\method{reveal}{default}(.invariant, .silent = TRUE, ...)

\method{reveal}{tbl_df}(.invariant, .silent = TRUE, ...)

\method{reveal}{xts}(.invariant, .silent = TRUE, ...)

\method{reveal}{matrix}(.invariant, .silent = TRUE, ...)
}
\arguments{
\item{.invariant}{A univariate or multivariate tabular (non-tidy) time-series.}

\item{.silent}{A logical. Should the optimization results be printed? The default is \code{TRUE}.}

\item{...}{Additional arguments the user may want to pass to \code{\link[ghyp]{fit.ghypuv}}.}
}
\value{
A \code{tibble} with the fitted parameters: 11 rows and 8 columns.
}
\description{
Performs a maximum likelihood estimation for univariate and
multivariate generalized hyperbolic distributions and reorder the fitted models
based on AIC information criterion. It's a wrapper around \code{\link[ghyp]{stepAIC.ghyp}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- colnames(EuStockMarkets)

reveal(x)
}
\references{
Pfaff, Bernhard. Financial risk modelling and portfolio optimization with R.
John Wiley & Sons, 2016.
}
\seealso{
\code{\link[ghyp]{stepAIC.ghyp}}
}
