% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_form.R
\name{scale_form}
\alias{scale_form}
\title{Fitting the form for scaling features}
\usage{
scale_form(dfr, remove.me = c(), exclude.binary = TRUE)
}
\arguments{
\item{remove.me}{Vector of variable names to be excluded from scaling}

\item{exclude.binary}{Boolean determining whether to exclude scaling binary features that are measured as numeric. Default is TRUE.}

\item{dat}{Dataframe with continuous variables intended for scaling}
}
\value{
Returns information about each feature needed for scaling.
}
\description{
Fitting the form for scaling features
}
\examples{
std.fit <- scale_form(iris, remove.me = c("Sepal.Width"))
iris2 <- scale_make(iris, std.fit, scaler = "standard")
head(iris2)
}
