% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_make.R
\name{scale_make}
\alias{scale_make}
\title{Making the scales on new data}
\usage{
scale_make(dfr, scale_data = trainScales, scaler = NA)
}
\arguments{
\item{scale_data}{The object created from calling `scale_form`}

\item{scaler}{Method for scaling. Options include: 'standard', 'minmax'}

\item{dat}{Dataframe with continuous variables intended for scaling (uses distribution from training data)}
}
\value{
Returns the original dataset `dat` with the desired numeric features scaled using the `scaler` method.
}
\description{
Making the scales on new data
}
\examples{
std.fit <- scale_form(iris, remove.me = c("Sepal.Width"))
iris2 <- scale_make(iris, std.fit, scaler = "standard")
head(iris2)
}
