% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.R
\name{uncover-marginal}
\alias{uncover-marginal}
\alias{new_marginal}
\title{Internal vctrs methods}
\usage{
new_marginal(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
No return value, called for side effects.
}
\description{
Internal vctrs methods
}
\keyword{internal}
