% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_normal}
\alias{fit_normal}
\alias{fit_normal.default}
\alias{fit_normal.tbl}
\alias{fit_normal.xts}
\alias{fit_normal.matrix}
\alias{fit_normal.numeric}
\title{Estimation of the Normal Distribution}
\usage{
fit_normal(x)

\method{fit_normal}{default}(x)

\method{fit_normal}{tbl}(x)

\method{fit_normal}{xts}(x)

\method{fit_normal}{matrix}(x)

\method{fit_normal}{numeric}(x)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs maximum likelihood estimation on the normal distribution
(univariate and multivariate). Wrappers \code{\link[ghyp]{fit.gaussuv}}
and \code{\link[ghyp]{fit.gaussmv}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_normal(x)

# univariate estimation
fit_normal(x[ , 4, drop = FALSE])
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}}
\code{\link{fit_vg}} \code{\link{fit_t}}
}
