% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.wrap.R
\name{print.IBIS}
\alias{print.IBIS}
\title{Print IBIS}
\usage{
\method{print}{IBIS}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"IBIS"}}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects
}
\description{
Prints summary information from an IBIS object.
}
\details{
When running the function \code{\link[=IBIS.logreg]{IBIS.logreg()}} the printed
information will contain information regarding; the number of samples, the
mean of those samples and the log Bayesian evidence of the posterior.
}
\seealso{
\code{\link[=IBIS.logreg]{IBIS.logreg()}}
}
\keyword{IBIS}
\keyword{print}
