% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNCOVER.R
\name{print.UNCOVER}
\alias{print.UNCOVER}
\title{Print UNCOVER}
\usage{
\method{print}{UNCOVER}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"UNCOVER"}}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects
}
\description{
Prints summary information from an UNCOVER object.
}
\details{
When running the function \code{\link[=UNCOVER]{UNCOVER()}} the printed
information will contain information regarding; the number of clusters, the
cluster sizes, the sub-model log Bayesian evidences and the total model log
Bayesian evidence.
}
\seealso{
\code{\link[=UNCOVER]{UNCOVER()}}
}
\keyword{UNCOVER}
\keyword{print}
